Implement feature for web browser opening up on its own upon running bench routes either through `make run` or through the executable.
This would take out the hassle from having to enter a url on
your own into the browser. Currently, the browser to be opened
has been stored as a variable, called browserName and has been
defaulted to firefox, but the ultimate goal is to be able to
pass the browser name as a flag and open a default if none is
provided.

But that might need command line parsing, one feature that is
in the pipeline after which we can come back and implement this
remaining functionality.

The changes here have been made keeping in mind that this code
must work on all the three intended operating systems namely
Linux, Windows and MacOS.